+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## v0.4.2 (2024-07-16)
+* Added Generator structure
+
 ## v0.4.1 (2024-07-05)
 * Added refreshThread method to Context